Added Gauner WIth Gun & Balancing Stuff

+++ Assets/Logic/Code/Utilities/FloatExtensions.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public static class FloatExtensions 
 {
-	public static bool IsNearlyEqual(this float a, float b, int epsilon)
+	public static bool IsNearlyEqual(this float a, float b, float epsilon)
 	{
 		return Ultra.Utilities.IsNearlyEqual(a, b, epsilon);
 	}
